OpenCL Typemaps (#886)
* Improve handling of enums that are empty in base spec

* Generator run

* Always remove extension tag from enum names

* Generator run

* Skip extension enums with no members to avoid useless name cluttering

* Skip extension enums only if core enum is present

* Generator run

* OpenCL typemaps

* Update OpenCL typemaps

* Make sure calling convention is carried through FunctionSignatureBuilder

* Post-rebase fix

* Handle nint-typed enums

* Move structs to corresponding extension projects

* Missed line

* Generator run

+++ src/OpenCL/Extensions/Silk.NET.OpenCL.Extensions.ARM/Enums/ImportProperties.gen.cs
@@ -10,7 +10,7 @@ using Silk.NET.Core.Attributes;
 namespace Silk.NET.OpenCL.Extensions.ARM
 {
     [NativeName("Name", "cl_import_properties_arm")]
-    public enum ImportProperties : int
+    public enum ImportProperties : ulong
     {
         [NativeName("Name", "CL_IMPORT_TYPE_ARM")]
         Type = 0x40B2,
